requirements.txt
anyio==3.7.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
idna==3.4
jwt==1.3.1
passlib==1.7.4
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
